requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# HTTP and API dependencies
requests>=2.31.0
httpx>=0.25.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Optional: Redis for session storage
redis>=5.0.0

# Optional: Data visualization helpers
matplotlib>=3.7.0
plotly>=5.17.0
